requirements.txt
alembic==1.15.1
annotated-types==0.7.0
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.39
SQLAlchemy-Utils==0.41.2
typing_extensions==4.12.2
Werkzeug==3.1.3
